Document field policy scope and scrub agg keys
